PATCH data completed

+++ weather/src/main/java/com/practice/weather/midTerm/land/repository/MidTermLandRepository.java
@@ -21,6 +21,6 @@ public interface MidTermLandRepository extends JpaRepository<MidTermLandEntity, 
     List<MidTermLandEntity> selectListByLocation(Pageable pageable, String location);
 
     @Query(value = "SELECT COUNT(1) FROM MID_TERM_LAND WHERE REG_ID = :location", nativeQuery = true)
-    int countByLocation(String location);
+    long countByLocation(String location);
     
 }
